apps/plans: move export from view to exports, slightly change export order
